Cleaned up old code and renamed scraping to scraper
